Update Issue 49 CostumeAdapter and CostumeActivity Refactoring done.
This issue is now on hold, because the current version of Paintroid won't let me reproduce the problem. Waiting for the cropping feature to be implemented.

+++ catroid/src/at/tugraz/ist/catroid/content/bricks/ChangeVolumeByBrick.java
@@ -109,7 +109,7 @@ public class ChangeVolumeByBrick implements Brick, OnClickListener {
 		dialog.setPositiveButton(context.getString(R.string.ok), new DialogInterface.OnClickListener() {
 			public void onClick(DialogInterface dialog, int which) {
 				try {
-					volume = Double.parseDouble(input.getText().toString());
+					volume = Float.parseFloat(input.getText().toString());
 				} catch (NumberFormatException exception) {
 					Toast.makeText(context, R.string.error_no_number_entered, Toast.LENGTH_SHORT).show();
 				}
